Deps: Remove squareup okhttp3 dependency from automattic tracks
This dependency was added as part of this
(9b00ec5a41dc6cb260ad8c937f9953131bd3bb62) commit, which added the
'CrashLoggingOkHttpInterceptorProvider' class on the project. But then,
as part of this (24ae7506517ea59a8c725fbaba222c07720fbcfc) other commit,
this 'CrashLoggingOkHttpInterceptorProvider' class got moved into the
newly created 'crashlogging' module. As such, this dependency is no
longer needed on 'AutomatticTracks'.

------------------------------------------------------------------------

This removal was suggested by the dependency analysis report, see below:

Advice for :AutomatticTracks
Unused dependencies which should be removed:
  implementation 'com.squareup.okhttp3:okhttp:4.9.0'