Build: Update dependency analysis configuration on sample tracks app
This 'unused' related advise for the 'sampletracksapp' module is
actually correct and 'AutomatticTracks' can be removed as a dependency
since it is actually unused and not being utilized from within the
sample tracks app.

However, I chose to keep it there and suppress the advise instead in
order to preserve the sample app default configuration, its purpose
being to test all those individual modules:
- AutomatticTracks
- crashlogging

As such this advice is instead suppressed and the 'onUnusedDependencies'
exclusion configuration were used to achieve that.

------------------------------------------------------------------------

Advice for :sampletracksapp
Unused dependencies which should be removed:
  implementation project(':AutomatticTracks')