Don't include `fetchedAt` as a part of `AssignmentsDto`
This property is not a part of contract with API. In previous commit I added this for convenience, but later I realised it'll be confusing (especially in the future) to mix API contract with requirements of experimentation module.

Hence this change, adding more logic to `FileBasedCache` but making things more readable.