Provide dispatchers and introduce `FileBasedCache#latest`
In order to remove `runBlocking` from `ExPlat`, we have to have a way to get value from cache in a way, that doesn't require sync file-operation. That's why I introduce `latest` field, which keeps the latest cache value in memory.

Providing dispatchers is considered a good practice and might come handy during tests