`ExPlat#getVariation` should not offer `shouldRefreshIfStale`
This is a confusing API, because, even if user requests a refresh, we don't return fresh data immediately, but only queue this to fetch. This is not something I'd expect as the SDK consumer.

Also, we already have specialized refresh methods. I don't think we need more.

I plan to remove `shouldRefreshIfStale` in another commit.